# docs/api/openapi.yaml
...
openapi: 3.1.0
...
info:
  title: API de Finanzas Inteligentes
  version: 1.0.0-draft
  description: Contrato ejecutable derivado de CONVENCIONES.md, ENDPOINTS.md, CONTRATOS.md y TRAZABILIDAD_RF.md.
servers:
- url: https://api.finanzas.example/api/v1
  description: Producción (referencial)
- url: http://localhost:8080/api/v1
  description: Desarrollo local
security:
- bearerAuth: []
...
paths:
...
  "/documentos-financieros":
    post:
      tags:
      - Documentos, OCR y XML SIFEN
      summary: POST /documentos-financieros
      operationId: postDocumentosFinancieros
      x-requisitos-funcionales:
      - RF-07
      - RF-08
      x-etapa: 5
      x-pruebas-contrato:
      - CT-RF-07
      - CT-RF-08
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - name: tipo
        in: query
        required: false
        schema:
          type: string
      - name: estado
        in: query
        required: false
        schema:
          type: string
      - name: desde
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: hasta
        in: query
        required: false
        schema:
          type: string
          format: date
      - "$ref": "#/components/parameters/IdempotencyKey"
      - name: X-Content-SHA256
        in: header
        required: true
        schema:
          type: string
          pattern: "^[0-9a-fA-F]{64}$"
      responses:
        '202':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
            Location:
              "$ref": "#/components/headers/Location"
            ETag:
              "$ref": "#/components/headers/ETag"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DocumentoFinancieroResponse"
        '400':
          "$ref": "#/components/responses/Problem400"
        '401':
          "$ref": "#/components/responses/Problem401"
        '403':
          "$ref": "#/components/responses/Problem403"
        '409':
          "$ref": "#/components/responses/Problem409"
        '413':
          "$ref": "#/components/responses/Problem413"
        '415':
          "$ref": "#/components/responses/Problem415"
        '422':
          "$ref": "#/components/responses/Problem422"
        '503':
          "$ref": "#/components/responses/Problem503"
        '500':
          "$ref": "#/components/responses/Problem500"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required:
              - archivo
              - tipo
              - ambito
              properties:
                archivo:
                  type: string
                  format: binary
                tipo:
                  type: string
                  enum:
                  - imagen
                  - pdf
                  - xml-sifen
                ambito:
                  type: string
                  enum:
                  - privado
                  - familiar
                grupoFamiliarId:
                  type:
                  - string
                  - 'null'
                  format: uuid
    get:
      tags:
      - Documentos, OCR y XML SIFEN
      summary: GET /documentos-financieros
      operationId: getDocumentosFinancieros
      x-requisitos-funcionales:
      - RF-07
      - RF-08
      x-etapa: 5
      x-pruebas-contrato:
      - CT-RF-07
      - CT-RF-08
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - "$ref": "#/components/parameters/Cursor"
      - "$ref": "#/components/parameters/Limite"
      - name: tipo
        in: query
        required: false
        schema:
          type: string
      - name: estado
        in: query
        required: false
        schema:
          type: string
      - name: desde
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: hasta
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        '200':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PaginaDocumentoFinancieroResponse"
        '400':
          "$ref": "#/components/responses/Problem400"
        '401':
          "$ref": "#/components/responses/Problem401"
        '500':
          "$ref": "#/components/responses/Problem500"
  "/documentos-financieros/{documentoId}":
    get:
      tags:
      - Documentos, OCR y XML SIFEN
      summary: GET /documentos-financieros/{documentoId}
      operationId: getDocumentosFinancierosByDocumentoId
      x-requisitos-funcionales:
      - RF-07
      - RF-08
      x-etapa: 5
      x-pruebas-contrato:
      - CT-RF-07
      - CT-RF-08
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - name: documentoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
            ETag:
              "$ref": "#/components/headers/ETag"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DocumentoFinancieroResponse"
        '401':
          "$ref": "#/components/responses/Problem401"
        '404':
          "$ref": "#/components/responses/Problem404"
        '500':
          "$ref": "#/components/responses/Problem500"
    delete:
      tags:
      - Documentos, OCR y XML SIFEN
      summary: DELETE /documentos-financieros/{documentoId}
      operationId: deleteDocumentosFinancierosByDocumentoId
      x-requisitos-funcionales:
      - RF-07
      - RF-08
      x-etapa: 5
      x-pruebas-contrato:
      - CT-RF-07
      - CT-RF-08
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - name: documentoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - "$ref": "#/components/parameters/IfMatch"
      responses:
        '204':
          description: Operación completada sin cuerpo.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
        '401':
          "$ref": "#/components/responses/Problem401"
        '404':
          "$ref": "#/components/responses/Problem404"
        '409':
          "$ref": "#/components/responses/Problem409"
        '412':
          "$ref": "#/components/responses/Problem412"
        '500':
          "$ref": "#/components/responses/Problem500"
  "/documentos-financieros/{documentoId}/descargas":
    post:
      tags:
      - Documentos, OCR y XML SIFEN
      summary: POST /documentos-financieros/{documentoId}/descargas
      operationId: postDocumentosFinancierosByDocumentoIdDescargas
      x-requisitos-funcionales:
      - RF-07
      - RF-08
      x-etapa: 5
      x-pruebas-contrato:
      - CT-RF-07
      - CT-RF-08
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - name: documentoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '201':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
            Location:
              "$ref": "#/components/headers/Location"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DescargaResponse"
        '401':
          "$ref": "#/components/responses/Problem401"
        '404':
          "$ref": "#/components/responses/Problem404"
        '409':
          "$ref": "#/components/responses/Problem409"
        '429':
          "$ref": "#/components/responses/Problem429"
        '503':
          "$ref": "#/components/responses/Problem503"
        '500':
          "$ref": "#/components/responses/Problem500"
  "/procesamientos-documentales/{procesamientoId}":
    get:
      tags:
      - Documentos, OCR y XML SIFEN
      summary: GET /procesamientos-documentales/{procesamientoId}
      operationId: getProcesamientosDocumentalesByProcesamientoId
      x-requisitos-funcionales:
      - RF-07
      - RF-08
      - RF-09
      x-etapa: 5
      x-pruebas-contrato:
      - CT-RF-07
      - CT-RF-08
      - CT-RF-09
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - name: procesamientoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
            ETag:
              "$ref": "#/components/headers/ETag"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ProcesamientoDocumentalResponse"
        '401':
          "$ref": "#/components/responses/Problem401"
        '404':
          "$ref": "#/components/responses/Problem404"
        '500':
          "$ref": "#/components/responses/Problem500"
    patch:
      tags:
      - Documentos, OCR y XML SIFEN
      summary: PATCH /procesamientos-documentales/{procesamientoId}
      operationId: patchProcesamientosDocumentalesByProcesamientoId
      x-requisitos-funcionales:
      - RF-07
      - RF-08
      - RF-09
      x-etapa: 5
      x-pruebas-contrato:
      - CT-RF-07
      - CT-RF-08
      - CT-RF-09
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - name: procesamientoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - "$ref": "#/components/parameters/IfMatch"
      responses:
        '200':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
            ETag:
              "$ref": "#/components/headers/ETag"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ProcesamientoDocumentalResponse"
        '400':
          "$ref": "#/components/responses/Problem400"
        '401':
          "$ref": "#/components/responses/Problem401"
        '404':
          "$ref": "#/components/responses/Problem404"
        '409':
          "$ref": "#/components/responses/Problem409"
        '412':
          "$ref": "#/components/responses/Problem412"
        '422':
          "$ref": "#/components/responses/Problem422"
        '500':
          "$ref": "#/components/responses/Problem500"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CorreccionProcesamientoDocumentalRequest"
  "/documentos-financieros/{documentoId}/procesamientos-documentales":
    post:
      tags:
      - Documentos, OCR y XML SIFEN
      summary: POST /documentos-financieros/{documentoId}/procesamientos-documentales
      operationId: postDocumentosFinancierosByDocumentoIdProcesamientosDocumentales
      x-requisitos-funcionales:
      - RF-07
      - RF-08
      - RF-09
      x-etapa: 5
      x-pruebas-contrato:
      - CT-RF-07
      - CT-RF-08
      - CT-RF-09
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - name: documentoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - "$ref": "#/components/parameters/IdempotencyKey"
      responses:
        '202':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
            Location:
              "$ref": "#/components/headers/Location"
            ETag:
              "$ref": "#/components/headers/ETag"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ProcesamientoDocumentalResponse"
        '400':
          "$ref": "#/components/responses/Problem400"
        '401':
          "$ref": "#/components/responses/Problem401"
        '403':
          "$ref": "#/components/responses/Problem403"
        '404':
          "$ref": "#/components/responses/Problem404"
        '409':
          "$ref": "#/components/responses/Problem409"
        '422':
          "$ref": "#/components/responses/Problem422"
        '429':
          "$ref": "#/components/responses/Problem429"
        '503':
          "$ref": "#/components/responses/Problem503"
        '500':
          "$ref": "#/components/responses/Problem500"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PostDocumentosFinancierosByDocumentoIdProcesamientosDocumentalesRequest"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 100
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 100
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        pattern: ^\".+\"$
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        maxLength: 500
    Limite:
      name: limite
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  headers:
    CorrelationId:
      description: Identificador de correlación de la solicitud.
      schema:
        type: string
    ETag:
      description: Versión opaca del recurso.
      schema:
        type: string
    Location:
      description: URI del recurso creado o aceptado.
      schema:
        type: string
        format: uri-reference
  schemas:
...
    DocumentoFinancieroResponse:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        nombreOriginal:
          type: string
          example: factura.pdf
        tipo:
          type: string
          example: pdf
        mimeType:
          type: string
          example: application/pdf
        tamanoBytes:
          type: integer
          format: int64
          example: 248120
        sha256:
          type: string
          example: hex-sha256
        estadoArchivo:
          type: string
          example: disponible
        estadoProcesamiento:
          type: string
          example: pendiente
        procesamientoId:
          type: string
          format: uuid
        creadoEn:
          type: string
          format: date-time
          example: '2026-07-22T18:30:00Z'
        version:
          type: integer
          format: int64
          example: 1
      required:
      - id
      - nombreOriginal
      - tipo
      - mimeType
      - tamanoBytes
      - sha256
      - estadoArchivo
      - estadoProcesamiento
      - procesamientoId
      - creadoEn
      - version
    ProcesamientoDocumentalResponse:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        documentoId:
          type: string
          format: uuid
        tipo:
          type: string
          example: ocr
        estado:
          type: string
          example: incompleto
        confianza:
          type: number
          format: double
          example: 0.72
          minimum: 0
          maximum: 1
        datosDetectados:
          type: object
          additionalProperties: false
          properties:
            monto:
              type: integer
              format: int64
              example: 285000
              minimum: 1
            fecha:
              type: string
              format: date
              example: '2026-07-22'
            comercio:
              type: string
              example: Supermercado San Roque
            categoriaSugeridaId:
              type: string
              format: uuid
            cdcSifen:
              type:
              - string
              - 'null'
          required:
          - monto
          - fecha
          - comercio
          - categoriaSugeridaId
        advertencias:
          type: array
          items:
            type: string
            example: La fecha posee baja confianza
        iniciadoEn:
          type: string
          format: date-time
          example: '2026-07-22T18:30:01Z'
        finalizadoEn:
          type: string
          format: date-time
          example: '2026-07-22T18:30:04Z'
        version:
          type: integer
          format: int64
          example: 2
      required:
      - id
      - documentoId
      - tipo
      - estado
      - confianza
      - datosDetectados
      - advertencias
      - iniciadoEn
      - finalizadoEn
      - version
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required:
      - type
      - title
      - status
      - codigo
      - correlationId
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type:
          - string
          - 'null'
        instance:
          type:
          - string
          - 'null'
        codigo:
          type: string
          pattern: "^[a-z0-9_]+$"
        correlationId:
          type: string
        errores:
          type:
          - object
          - 'null'
          additionalProperties:
            type: array
            items:
              type: string
    Paginacion:
      type: object
      additionalProperties: false
      required:
      - siguienteCursor
      - hayMas
      - limite
      properties:
        siguienteCursor:
          type:
          - string
          - 'null'
        hayMas:
          type: boolean
        limite:
          type: integer
          minimum: 1
          maximum: 100
    DescargaResponse:
      type: object
      additionalProperties: false
      required:
      - url
      - expiraEn
      properties:
        url:
          type: string
          format: uri
        expiraEn:
          type: string
          format: date-time
...
    PaginaDocumentoFinancieroResponse:
      type: object
      additionalProperties: false
      required:
      - datos
      - paginacion
      properties:
        datos:
          type: array
          items:
            "$ref": "#/components/schemas/DocumentoFinancieroResponse"
        paginacion:
          "$ref": "#/components/schemas/Paginacion"
...
    PostDocumentosFinancierosByDocumentoIdProcesamientosDocumentalesRequest:
      type: object
      additionalProperties: false
      properties:
        tipo:
          type: string
          enum:
          - ocr
          - sifen
      minProperties: 1
      required:
      - tipo
    CorreccionProcesamientoDocumentalRequest:
      type: object
      additionalProperties: false
      required:
      - datosDetectados
      properties:
        datosDetectados:
          type: object
          minProperties: 1
          additionalProperties: true
...
  responses:
    Problem400:
      description: Solicitud o parámetros inválidos.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem401:
      description: Autenticación ausente o inválida.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem403:
      description: Permiso insuficiente.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem404:
      description: Recurso inexistente o no visible.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem409:
      description: Conflicto de estado o duplicado.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
...
    Problem412:
      description: ETag desactualizado.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem413:
      description: Contenido demasiado grande.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem415:
      description: Tipo de contenido no soportado.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem422:
      description: Regla de negocio no satisfecha.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem429:
      description: Límite de solicitudes excedido.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem500:
      description: Error interno no controlado.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem503:
      description: Dependencia temporalmente no disponible.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"